[Refactor] rank가 statistics안에 들어가도록 변경
